refactor(types): relocate claim value-key aliases to core leaf
Move ClaimValueKey, IdentityPartName and the new ColumnName alias into
apps.core.types so they sit in the dependency-free leaf rather than in
provenance.types / provenance.models.claim. This lets the claim-control
model bases reference them while importing only apps.core.

RelationshipClaimValue stays in provenance.types and IdentityPartValue in
claim.py, each now importing the relocated alias. ColumnNames in the
resolution engine narrows its element type to the new ColumnName.
