+++ backend/apps/provenance/models/__init__.py
@@ -14,7 +14,6 @@ from .citation_instance import CITATION_INSTANCE_LOCATOR_MAX_LENGTH, CitationIns
 from .claim import (
     Claim,
     ExistingClaimRow,
-    IdentityPartName,
     IdentityPartValue,
     make_claim_key,
 )
