+++ backend/apps/provenance/types.py
@@ -8,6 +8,11 @@ from __future__ import annotations
 
 from collections.abc import Mapping
 
+# ClaimValueKey lives in apps.core.types (the dependency-free leaf); imported here
+# only because RelationshipClaimValue below is defined in terms of it. Do not
+# remove this import — RelationshipClaimValue needs it in scope.
+from apps.core.types import ClaimValueKey
+
 type IngestSourceId = int
 """The primary key of an Ingest Source — the non-human Actor, such as the data
 patch pipeline or a bot, to which a claim may be attributed."""
@@ -19,11 +24,6 @@ claim was written or retracted."""
 type IngestRunId = int
 """The primary key of an IngestRun — the bulk import to which a change belongs,
 which marks it as ingested rather than interactively created by a human."""
-
-type ClaimValueKey = str
-"""Names one slot inside a relationship claim's value — the person and role of a
-credit, the count of a gameplay feature. The schema fixes which slots form member
-identity and which carry payload."""
 
 type RelationshipClaimValue = Mapping[ClaimValueKey, object]
 """A relationship claim's stored value payload: the namespace-specific slots its
